+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Given a version number `MAJOR.MINOR.PATCH`, we increment the:
 ## Unreleased
 
 - Fix `--cloud-status` flag when stacks were synchronized with uppercase in the `stack.id`.
+- Fix `terramate cloud drift show` when used in stacks containing uppercase in the `stack.id`.
 
 ## v0.5.3